Fixed bugs related guider reseting workflow + implemented proper ASI Camera class
